wensday

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -15,39 +15,58 @@ class CategoryController extends Controller
 {
     public function childs(Request $request)
     {
-
-        Category::find($request->category_id);
-        $categories = SubCategory::
-        where('language_id', language())
-            ->where('category_id', $request->category_id)
-            ->orderBy('sort')
-            ->get();
-        return response()->json($categories);
+        if(auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            Category::find($request->category_id);
+            $categories = SubCategory::
+            where('language_id', language())
+                ->where('category_id', $request->category_id)
+                ->orderBy('sort')
+                ->get();
+            return response()->json($categories);
+        }else{
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
     }
 
 
     public function index()
     {
-        $data['categories'] = Category::select('id', 'name', 'language_id', 'featured', 'show_in_footer', 'sort')
-            ->where('language_id', language())
-            ->orderBy('sort')
-            ->get();
-        return view('admin.category.index')->with($data);
+        if(auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            $data['categories'] = Category::select('id', 'name', 'language_id', 'featured', 'show_in_footer', 'sort')
+                ->where('language_id', language())
+                ->orderBy('sort')
+                ->get();
+            return view('admin.category.index')->with($data);
+        }else{
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
     }
 
 
     // return view of adding data
     public function add()
     {
-        $data['countries'] = Country::where('language_id', language())->get();
+        if(auth()->guard('admin')->user()->hasPermissionTo('add categories')){
+            $data['countries'] = Country::where('language_id', language())->get();
 
-        return view('admin.category.add')->with($data);
+            return view('admin.category.add')->with($data);
+        }else{
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
     }
 
 
     // storing data in db
     public function store(CategoryRequest $request)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('add categories')){
+        
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         //dd($request->all());
         $data = $request->validated();
         $data['language_id'] = language();
@@ -82,6 +101,10 @@ class CategoryController extends Controller
     // return view of editing data
     public function edit($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('edit categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $data['countries'] = Country::where('language_id', language())->get();
         $data['row'] = Category::findOrFail($id);
         return view('admin.category.edit')->with($data);
@@ -91,6 +114,10 @@ class CategoryController extends Controller
     // storing data in db
     public function update(CategoryRequest $request)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('edit categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $data = $request->validated();
         $old = Category::findOrFail($request->id);
 
@@ -139,6 +166,10 @@ class CategoryController extends Controller
     // deleteing row from db  ( soft delete )
     public function delete($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('delete categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         Category::findOrFail($id)->delete();
         session()->flash('message', trans('site.deleted_success'));
         return back();
@@ -148,6 +179,10 @@ class CategoryController extends Controller
 
     public function showInFooter($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $category = Category::findOrFail($id);
         $category->show_in_footer = 'yes';
         $category->save();
@@ -158,6 +193,10 @@ class CategoryController extends Controller
 
     public function hideFromFooter($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $category = Category::findOrFail($id);
         $category->show_in_footer = 'no';
         $category->save();
@@ -168,6 +207,10 @@ class CategoryController extends Controller
 
     public function showInHome($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $category = Category::findOrFail($id);
         $category->featured = 'yes';
         $category->save();
@@ -178,6 +221,10 @@ class CategoryController extends Controller
 
     public function hideFromHome($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $category = Category::findOrFail($id);
         $category->featured = 'no';
         $category->save();
@@ -189,6 +236,10 @@ class CategoryController extends Controller
     // delete mutli row from table
     public function deleteMulti(Request $request)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('delete categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         foreach ($request->deleteMulti as $value) {
             Category::findOrFail($value)->delete();
 
@@ -202,6 +253,10 @@ class CategoryController extends Controller
     // delete mutli row from table
     public function sort(Request $request)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $i = 1;
         foreach ($request->sort as $value) {
             $data = Category::find($value);
@@ -217,6 +272,10 @@ class CategoryController extends Controller
     // visibility  for this item ( active or not active  -- change featured of this item )
     public function visibility($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $data = Category::findOrFail($id);
 
         switch ($data->featured) {
@@ -245,6 +304,10 @@ class CategoryController extends Controller
     // return view of seo data
     public function seo($id)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $data['seo'] = Category::where('language_id', language())->findOrFail($id);
         $data['seoData'] = json_decode($data['seo']['seo']);
         return view('admin.category.seo')->with($data);
@@ -255,6 +318,10 @@ class CategoryController extends Controller
     // storing data in db
     public function updateSeo(Request $request)
     {
+        if(!auth()->guard('admin')->user()->hasPermissionTo('read categories')){
+            session()->flash('error_permission',trans('site.added_success'));
+            return back();
+        }
         $data = $request->except('id', '_token');
         // updating data in db
         $seoData = Category::find($request->id);
